requirements.txt
pyzmq==22.0.3
redis==3.5.3
